Add segmentation tool example
The name of the tool class is crucial. To appear in the tool selection
box of the segmentation views, the class name must end in either
SegTool2D or SegTool3D.

If an own state machine is provided, is it assumed that there is also
an event config file with the same name, suffixed with Config.xml.
For example, Paint.xml must be accompanied by PaintConfig.xml.

To guarantee the appearance of the tool in the segmentation views,
we must assure to load the module before a segmentation view is opened.
Otherwise the tool is not initialized in time and cannot be found by the
ITK object factory mechanism used in segmentation tools. For this purpose,
a dummy function was added to the module and an extra plugin with eager
activation policy was added. The sole task of this plugin is to call the
dummy function, forcing the runtime to load the module. Eager plugins are
loaded very early in MITK before any other plugin.